requirements.txt
Flask==2.3.3
Pillow==10.0.1
pyzbar==0.1.9
opencv-python==4.8.1.78
numpy==1.24.3
zxing-cpp==2.2.1